+++ Admin/User/UserAddressPage.php
@@ -9,22 +9,28 @@ declare(strict_types=1);
 
 namespace OxidEsales\Codeception\Admin\User;
 
-use OxidEsales\Codeception\Admin\Component\AdminUserAddressesForm;
-use OxidEsales\Codeception\Page\Page;
 use OxidEsales\Codeception\Admin\DataObject\AdminUserAddresses;
+use OxidEsales\Codeception\Module\Translation\Translator;
+use OxidEsales\Codeception\Page\Page;
 
-/**
- * Class Users
- *
- * @package OxidEsales\Codeception\Admin
- */
 class UserAddressPage extends Page
 {
     use UserList;
-    use AdminUserAddressesForm;
 
     public $deleteAddressInput = "//input[@value='Delete']";
     public $addressesTabAddressSelect = 'oxaddressid';
+    public string $addressTitleField = "//select[@name='editval[oxaddress__oxsal]']";
+    public string $addressFirstNameField = "//input[@name='editval[oxaddress__oxfname]']";
+    public string $addressLastNameField = "//input[@name='editval[oxaddress__oxlname]']";
+    public string $addressCompanyField = "//input[@name='editval[oxaddress__oxcompany]']";
+    public string $addressStreetField = "//input[@name='editval[oxaddress__oxstreet]']";
+    public string $addressStreetNumberField = "//input[@name='editval[oxaddress__oxstreetnr]']";
+    public string $addressZipCodeField = "//input[@name='editval[oxaddress__oxzip]']";
+    public string $addressCityField = "//input[@name='editval[oxaddress__oxcity]']";
+    public string $addressAdditionalInformationField = "//input[@name='editval[oxaddress__oxaddinfo]']";
+    public string $addressCountryIdField = "//select[@name='editval[oxaddress__oxcountryid]']";
+    public string $addressPhoneField = "//input[@name='editval[oxaddress__oxfon]']";
+    public string $addressFaxField = "//input[@name='editval[oxaddress__oxfax]']";
 
     /**
      * @return $this
@@ -55,6 +61,33 @@ class UserAddressPage extends Page
     }
 
     /**
+     * @param AdminUserAddresses $adminUserAddresses
+     * @return $this
+     */
+    public function editUserAddress(AdminUserAddresses $adminUserAddresses): self
+    {
+        $I = $this->user;
+
+        $I->selectOption($this->addressTitleField, $adminUserAddresses->getTitle());
+        $I->fillField($this->addressFirstNameField, $adminUserAddresses->getFirstName());
+        $I->fillField($this->addressLastNameField, $adminUserAddresses->getLastName());
+        $I->fillField($this->addressCompanyField, $adminUserAddresses->getCompany());
+        $I->fillField($this->addressStreetField, $adminUserAddresses->getStreet());
+        $I->fillField($this->addressStreetNumberField, $adminUserAddresses->getStreetNumber());
+        $I->fillField($this->addressZipCodeField, $adminUserAddresses->getZip());
+        $I->fillField($this->addressCityField, $adminUserAddresses->getCity());
+        $I->fillField($this->addressAdditionalInformationField, $adminUserAddresses->getAdditionalInfo());
+        $I->selectOption($this->addressCountryIdField, $adminUserAddresses->getCountryId());
+        $I->fillField($this->addressPhoneField, $adminUserAddresses->getPhone());
+        $I->fillField($this->addressFaxField, $adminUserAddresses->getFax());
+
+        $I->click(Translator::translate('GENERAL_SAVE'));
+        $I->waitForDocumentReadyState();
+
+        return $this;
+    }
+
+    /**
      * @param AdminUserAddresses $adminUserAddress
      * @return $this
      */
@@ -75,7 +108,7 @@ class UserAddressPage extends Page
         $I->seeInField($this->addressZipCodeField, $adminUserAddress->getZip());
         $I->seeInField($this->addressCityField, $adminUserAddress->getCity());
         $I->seeOptionIsSelected($this->addressCountryIdField, $adminUserAddress->getCountryId());
-        $I->seeInField($this->addressAdditonalInformationField, $adminUserAddress->getAdditionalInfo());
+        $I->seeInField($this->addressAdditionalInformationField, $adminUserAddress->getAdditionalInfo());
         $I->seeInField($this->addressPhoneField, $adminUserAddress->getPhone());
         $I->seeInField($this->addressFaxField, $adminUserAddress->getFax());